feat: enhance full name validation in UserForm and refactor UserTable for improved user management
